Huge refactoring of pybullet environment
- main class simplified, index utile broken into child classes
- updated environments to use such child classes
