+++ mushroom_rl/environments/pybullet_envs/locomotion/hopper.py
@@ -29,13 +29,27 @@ if __name__ == '__main__':
     from mushroom_rl.core import Core, Agent, Environment
     from mushroom_rl.utils.dataset import compute_J
 
+    np.random.seed(1)
+
+    def step_callback(dataset):
+        step = dataset[0]
+        action = step[1]
+        reward = step[2]
+        state = step[3]
+        last = step[-1]
+
+        print('action ', action)
+        print('state ', state)
+        print('reward ', reward)
+        print(last)
+
     class DummyAgent(Agent):
         def __init__(self, n_actions):
             self._n_actions = n_actions
 
         def draw_action(self, state):
             time.sleep(1/60)
-
+            #return 1e-3*np.ones(self._n_actions)
             return np.random.randn(self._n_actions)
 
         def episode_start(self):
@@ -49,8 +63,8 @@ if __name__ == '__main__':
 
     agent = DummyAgent(mdp.info.action_space.shape[0])
 
-    core = Core(agent, mdp)
-    dataset = core.evaluate(n_episodes=10, render=False)
+    core = Core(agent, mdp, callback_step=step_callback)
+    dataset = core.evaluate(n_episodes=10, render=False, quiet=True)
     print('reward: ', compute_J(dataset, mdp.info.gamma))
     print("mdp_info state shape", mdp.info.observation_space.shape)
     print("actual state shape", dataset[0][0].shape)
