+++ mushroom_rl/utils/pybullet/__init__.py
@@ -1,0 +1,4 @@
+from .observation import PyBulletObservationType
+from .index_map import IndexMap
+from .viewer import PyBulletViewer
+from .joints_helper import JointsHelper
